Fibonacci from long long to long

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+
 /**
  *main - prints the first 50 fibonacci numbers starting with 1 and 2
  *
@@ -7,9 +8,9 @@
 int main()
 {
     int count;
-    long long int f1 = 1;
-    long long int f2 = 2;
-    long long int next;
+    long int f1 = 1;
+    long int f2 = 2;
+    long int next;
     int flimit = 50;
 
     for (count = 2; count <= flimit; count++)
@@ -17,7 +18,7 @@ int main()
         next = f1 + f2;
         f1 = f2;
         f2 = next;
-        printf("%llu", next);
+        printf("%lu", next);
         if (count < flimit)
             printf(", ");
     }
